Move individualPage.js to screen dir & rename it to individualPageScreen.js
