+++ screens/individualPageScreen.js
